+++ middleware/headers.js
@@ -54,7 +54,7 @@ var middleware = module.exports = {
   },
   // Build the "Allow" response header
   allow: function (request, response, next) {
-    var allowed = request.app.activeVerbs().map(function (verb) {
+    var allowed = request.baucis.controller.activeVerbs().map(function (verb) {
       return verb.toUpperCase();
     });
 
@@ -63,7 +63,12 @@ var middleware = module.exports = {
   },
   // Build the "Accept" response header
   accept: function (request, response, next) {
-    response.set('Accept', 'application/json');
+    var putOff = (request.baucis.controller.get('put') === false);
+    var postOff = (request.baucis.controller.get('post') === false);
+
+    if (putOff && postOff) return next();
+
+    response.set('Accept', 'application/json, application/x-www-form-urlencoded');
     next();
   },
   // Add the "Location" response header