+++ CHANGES.md
@@ -1,16 +1,51 @@
 Baucis Change Log
 =================
 
+v0.6.27
+-------
+Fix an issue where POST middleware could be called at query stage
+
+v0.6.26
+-------
+Fix tests for latest version of Mongoose & Express
+
+v0.6.25
+-------
+Respond with 400 for requests with malformed IDs.  Respond with 405 for disabled verbs.
+
+v0.6.24
+-------
+Adds default support for form POST (`application/x-www-form-urlencoded`).  Also allows use of default Express request query string parser (`conditions[name]=Tomato`), in addition to the JSON string format that is supported.
+
+v0.6.23
+-------
+Adds `request.baucis.controller`.  This also fixes issue #76 that was occurring with newer Express versions when they were used as a peer dependency.
+
+v0.6.22
+-------
+Move Express to peer dependency.
+
+v0.6.21
+-------
+
+Fix an issue where the `Location` response header would be set incorrectly.  Special thanks to feychenie.
+
+v0.6.20
+-------
+
+Fix an issue with swagger when using models with hyphenated names.  Special
+thanks to feychenie.
+
 v0.6.19
 -------
 
 Better error handling when detecing selected fields when populating via query.
 
-v0.6.18 
+v0.6.18
 -------
 
-Fixes issue #72.  The `Location` header is now only set for POST responses.  
-This fixes the related issue of response headers being too large when doing 
+Fixes issue #72.  The `Location` header is now only set for POST responses.
+This fixes the related issue of response headers being too large when doing
 GET requests that return a large number of documents.
 
 v0.6.17
@@ -102,7 +137,7 @@ Added X-Baucis-Push header for PUTs
  * Feature must be enabled per-controller
  * Fields must be whitelisted by setting the 'allow push' controller option, e.g.
 
-    var controller = baucis.rest({ 
-      singular: 'arbitrary', 
-      'allow push': 'field some.path some.other.path' 
+    var controller = baucis.rest({
+      singular: 'arbitrary',
+      'allow push': 'field some.path some.other.path'
     });